밸록스 이름 변경

+++ 3D_RPG_Project/Assets/_3D RPG/Scripts/Enemy/BossUi.cs
@@ -31,6 +31,6 @@ public class BossUi : MonoBehaviour
         }
 
 
-        hpBar.value = _BossStatus.GetCurrentHp() / (float)_BossStatus.GetMaxHp();
+        hpBar.value = (float)_BossStatus.GetCurrentHp() / (float)_BossStatus.GetMaxHp();
     }
 }
